# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FleetSaaS school bus tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List emergency notices for current tenant
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an emergency announcement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── STATIONS ───────────────────────────────────────────────────────────────
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List all drivers for this tenant
      responses:
        "200":
          description: Driver list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Add a new driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    patch:
      operationId: patchDriver
      tags: [drivers]
      summary: Update driver details (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    delete:
      operationId: deleteDriver
      tags: [drivers]
      summary: Remove a driver (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List all routes for this tenant
      responses:
        "200":
          description: Route list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a new route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    patch:
      operationId: updateRoute
      tags: [routes]
      summary: Update route name or assigned driver/vehicle (fleet swap)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteUpdate"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /geocode:
    get:
      operationId: geocodeAddress
      tags: [geocode]
      summary: Look up coordinates for a station name or address (Nominatim)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Geocode results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeocodeResult"

  # ─── SUBSCRIPTIONS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Announcement:
      type: object
      required: [id, tenantId, message, severity, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        message:
          type: string
        messageNe:
          type: ["string", "null"]
        severity:
          type: string
          enum: [info, warning, emergency]
        createdAt:
          type: string
          format: date-time
    AnnouncementInput:
      type: object
      required: [message, severity]
      properties:
        message:
          type: string
        messageNe:
          type: string
        severity:
          type: string
          enum: [info, warning, emergency]
    Route:
      type: object
      required: [id, tenantId, name, isActive]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        driverName:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
    RouteInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        driverId:
          type: integer
        vehicleId:
          type: integer
    RouteUpdate:
      type: object
      properties:
        name:
          type: string
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        isActive:
          type: boolean
...
    GeocodeResult:
      type: object
      required: [displayName, lat, lng]
      properties:
        displayName:
          type: string
        lat:
          type: number
        lng:
          type: number
...
    Driver:
      type: object
      required: [id, name, phone, vehicleNumber]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        photoUrl:
          type: ["string", "null"]
        vehicleNumber:
          type: string
        isActive:
          type: boolean
    DriverInput:
      type: object
      required: [name, phone, vehicleNumber]
      properties:
        name:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
        vehicleNumber:
          type: string
    DriverUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        vehicleNumber:
          type: string
        photoUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
...